[don't backport to main] fix: topbar badges bg color on cloud (uses old menus style) (#6332)
## Summary

The topbar badges and cloud badges were changed to work with the new
menu system because it was developed on main - but on the cloud RC
branch, the old menu system still presides which uses a different topbar
background color. This PR fixes the badges to align with that.

┆Issue is synchronized with this [Notion
page](https://www.notion.so/PR-6332-don-t-backport-to-main-fix-topbar-badges-bg-color-on-cloud-uses-old-menus-style-2996d73d365081328f61f1e0fccbbbe5)
by [Unito](https://www.unito.io)

---------

Co-authored-by: GitHub Action <action@github.com>